tests/requirements.txt
pytest>=7.0.0
pytest-cov>=3.0.0
requests>=2.28.0
python-dotenv>=0.21.0
pytest-asyncio>=0.20.0
pytest-html>=3.2.0
pytest-xdist>=3.0.0
httpx>=0.23.0
minio>=7.0.0
pytest-docker>=1.0.0
